Aggiunto casting

+++ src/main/java/it/bigdatalab/applications/GroundTruths.java
@@ -100,7 +100,8 @@ public class GroundTruths {
         totalTime = endTime - startTime;
         System.out.println("Somma di tutte le distanze = "+ avg_distance);
         System.out.println("Numero di nodi nel grafo = "+ mGraph.numNodes());
-        double total_avg_distance =   ((double) avg_distance / (mGraph.numNodes()*(mGraph.numNodes()-1)));
+
+        double total_avg_distance =   ((double) avg_distance / ((double) mGraph.numNodes()*((double) (mGraph.numNodes()-1))));
         bfsResults.put("numNodes",mGraph.numNodes());
         bfsResults.put("numArcs",mGraph.numArcs());
         bfsResults.put("sum_distances",avg_distance);
